+++ data/common/export.py
@@ -2,7 +2,6 @@ import functools
 import json
 import logging
 import urllib.parse
-from copy import deepcopy
 from os.path import dirname
 
 import bw2calc
@@ -14,13 +13,22 @@ import requests
 from bw2io.utils import activity_hash
 from frozendict import frozendict
 
-from .impacts import bytrigram, main_method
+from . import (
+    bytrigram,
+    normalization_factors,
+    spproject,
+    with_corrected_impacts,
+    with_subimpacts,
+)
+from .impacts import main_method
 
 logging.basicConfig(level=logging.ERROR)
 
 PROJECT_ROOT_DIR = dirname(dirname(dirname(__file__)))
-IMPACTS_FILE = f"{PROJECT_ROOT_DIR}/public/data/impacts.json"
 COMPARED_IMPACTS_FILE = "compared_impacts.csv"
+
+with open(f"{PROJECT_ROOT_DIR}/public/data/impacts.json") as f:
+    IMPACTS_JSON = json.load(f)
 
 
 def check_ids(ingredients):
@@ -35,117 +43,8 @@ def check_ids(ingredients):
             )
 
 
-def compute_normalization_factors(impact_defs):
-    normalization_factors = {}
-    for k, v in impact_defs.items():
-        if v["ecoscore"]:
-            normalization_factors[k] = (
-                v["ecoscore"]["weighting"] / v["ecoscore"]["normalization"]
-            )
-        else:
-            normalization_factors[k] = 0
-    return normalization_factors
-
-
-def spproject(activity):
-    """return the current simapro project for an activity"""
-    match activity.get("database"):
-        case "Ginko":
-            return "Ginko w/o azadirachtin"
-        case "Ecobalyse":
-            # return a non existing project to force looking at brightway
-            return "EcobalyseIsNotASimaProProject"
-        case "Ecoinvent 3.9.1":
-            return "ADEME UPR"
-        case _:
-            return "AGB3.1.1 2023-03-06"
-
-
-def remove_detailed_impacts(processes):
-    result = list()
-    for process in processes:
-        new_process = deepcopy(process)
-        for k in new_process["impacts"].keys():
-            if k not in ("pef", "ecs"):
-                new_process["impacts"][k] = 0
-        result.append(new_process)
-    return result
-
-
-def order_json(data, filename):
-    """
-    Export data to a JSON file, with added newline at the end.
-    Make sure to sort impacts in the json file
-    """
-    if isinstance(data, list):
-        sorted_data = [
-            {**item, "impacts": sort_impacts(item["impacts"])}
-            if "impacts" in item
-            else item
-            for item in data
-        ]
-    elif isinstance(data, dict):
-        sorted_data = {
-            key: {**value, "impacts": sort_impacts(value["impacts"])}
-            if "impacts" in value
-            else value
-            for key, value in data.items()
-        }
-    else:
-        sorted_data = data
-    return sorted_data
-
-
-def sort_impacts(impacts):
-    # Define the desired order of impact keys
-    impact_order = [
-        "acd",
-        "cch",
-        "etf",
-        "etf-c",
-        "fru",
-        "fwe",
-        "htc",
-        "htc-c",
-        "htn",
-        "htn-c",
-        "ior",
-        "ldu",
-        "mru",
-        "ozd",
-        "pco",
-        "pma",
-        "swe",
-        "tre",
-        "wtu",
-        "pef",
-        "ecs",
-    ]
-    return {key: impacts[key] for key in impact_order if key in impacts}
-
-
 def progress_bar(index, total):
     print(f"Export in progress: {str(index)}/{total}", end="\r")
-
-
-def with_subimpacts(impacts):
-    """compute subimpacts"""
-    if not impacts:
-        return impacts
-    # etf-o = etf-o1 + etf-o2
-    impacts["etf-o"] = impacts["etf-o1"] + impacts["etf-o2"]
-    del impacts["etf-o1"]
-    del impacts["etf-o2"]
-    # etf = etf1 + etf2
-    impacts["etf"] = impacts["etf1"] + impacts["etf2"]
-    del impacts["etf1"]
-    del impacts["etf2"]
-    return impacts
-
-
-@functools.cache
-def cached_search(dbname, name, excluded_term=None):
-    return search(dbname, name, excluded_term)
 
 
 def search(dbname, name, excluded_term=None):
@@ -154,71 +53,6 @@ def search(dbname, name, excluded_term=None):
         results = [res for res in results if excluded_term not in res["name"]]
     assert len(results) >= 1, f"'{name}' was not found in Brightway"
     return results[0]
-
-
-def with_corrected_impacts(impacts_ecobalyse, processes_fd, impacts_key="impacts"):
-    """Add corrected impacts to the processes"""
-    corrections = {
-        k: v["correction"] for (k, v) in impacts_ecobalyse.items() if "correction" in v
-    }
-    processes = dict(processes_fd)
-    processes_updated = {}
-    for key, process in processes.items():
-        # compute corrected impacts
-        for impact_to_correct, correction in corrections.items():
-            corrected_impact = 0
-            for correction_item in correction:  # For each sub-impact and its weighting
-                sub_impact_name = correction_item["sub-impact"]
-                if sub_impact_name in process[impacts_key]:
-                    sub_impact = process[impacts_key].get(sub_impact_name, 1)
-                    corrected_impact += sub_impact * correction_item["weighting"]
-                    del process[impacts_key][sub_impact_name]
-            process[impacts_key][impact_to_correct] = corrected_impact
-        processes_updated[key] = process
-    return frozendict(processes_updated)
-
-
-def with_aggregated_impacts(impacts_ecobalyse, processes_fd, impacts_key="impacts"):
-    """Add aggregated impacts to the processes"""
-
-    # Pre-compute normalization factors
-    normalization_factors = {
-        "ecs": {
-            k: v["ecoscore"]["weighting"] / v["ecoscore"]["normalization"]
-            for k, v in impacts_ecobalyse.items()
-            if v["ecoscore"] is not None
-        },
-        "pef": {
-            k: v["pef"]["weighting"] / v["pef"]["normalization"]
-            for k, v in impacts_ecobalyse.items()
-            if v["pef"] is not None
-        },
-    }
-
-    processes_updated = {}
-    for key, process in processes_fd.items():
-        updated_process = dict(process)
-        updated_impacts = updated_process[impacts_key].copy()
-
-        updated_impacts["pef"] = calculate_aggregate(
-            updated_impacts, normalization_factors["pef"]
-        )
-        updated_impacts["ecs"] = calculate_aggregate(
-            updated_impacts, normalization_factors["ecs"]
-        )
-
-        updated_process[impacts_key] = updated_impacts
-        processes_updated[key] = updated_process
-
-    return frozendict(processes_updated)
-
-
-def calculate_aggregate(process_impacts, normalization_factors):
-    # We multiply by 10**6 to get the result in µPts
-    return sum(
-        10**6 * process_impacts.get(impact, 0) * normalization_factors.get(impact, 0)
-        for impact in normalization_factors
-    )
 
 
 def display_changes(key, oldprocesses, processes):
@@ -449,14 +283,10 @@ def compare_impacts(frozen_processes, default_db, impact_defs):
 
 def plot_impacts(ingredient_name, impacts_smp, impacts_bw, folder, impact_defs):
     impact_labels = impacts_smp.keys()
-    normalization_factors = compute_normalization_factors(impact_defs)
+    nf = normalization_factors(impact_defs)
 
-    simapro_values = [
-        impacts_smp[label] * normalization_factors[label] for label in impact_labels
-    ]
-    brightway_values = [
-        impacts_bw[label] * normalization_factors[label] for label in impact_labels
-    ]
+    simapro_values = [impacts_smp[label] * nf[label] for label in impact_labels]
+    brightway_values = [impacts_bw[label] * nf[label] for label in impact_labels]
 
     x = numpy.arange(len(impact_labels))
     width = 0.35
@@ -516,6 +346,11 @@ def load_json(filename):
     """
     with open(filename, "r") as file:
         return json.load(file)
+
+
+@functools.cache
+def cached_search(dbname, name, excluded_term=None):
+    return search(dbname, name, excluded_term)
 
 
 def find_id(dbname, activity):
